Add ArlasToolkitSharedModule
